Add blocked settings and clean ups
