--- lib/rule_blocklist.js
@@ -1,15 +1,0 @@
-module.exports = [
-  "import/extensions",
-  "import/no-restricted-paths",
-  "import/no-unresolved",
-  "import/no-extraneous-dependencies",
-  "import/no-named-as-default",
-  "import/namespace",
-  "import/named",
-  "import/no-absolute-path",
-  "import/no-duplicates",
-  "import/no-named-as-default-member",
-  "import/no-cycle",
-  "node/no-hide-code-modules",
-  "node/no-missing-require",
-]
